+++ backend/src/main/java/io/ssafy/p/j14c103/homerun/domain/world/housing/RealEstatePropertyRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface RealEstatePropertyRepository extends JpaRepository<RealEstateProperty, Integer> {
+public interface RealEstatePropertyRepository extends JpaRepository<RealEstateProperty, Long> {
 
     Optional<RealEstateProperty> findByProviderId(String providerId);
